requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
Werkzeug==3.0.1
APScheduler==3.10.4
twilio==8.12.0
python-dotenv==1.0.1
gunicorn==21.2.0
